requirements.txt
Flask
mysql-connector-python
reportlab
Pillow
gunicorn